requirements.txt
aniso8601==9.0.1
blinker==1.6.2
build==0.10.0
click==8.1.3
coverage==7.3.2
delocate==0.10.4
Flask==2.3.2
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
gunicorn==21.2.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
pyelftools==0.29
PyJWT==2.8.0
pyproject_hooks==1.0.0
pytest==7.4.2
pytest-env==1.1.0
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.3
